requirements.txt
opencv-python
mediapipe
numpy
pillow
scikit-image
